fix with swagger

+++ routes/api.php
@@ -32,7 +32,7 @@ Route::post('/login', [AuthController::class, 'login']);
 Route::post('/validate-token', [AuthController::class, 'validateToken']);
 Route::post('/forgot-password', [AuthController::class, 'forgotPassword']);
 
-Route::post('/reset-password', [AuthController::class,'resetPassword']);
+Route::post('/reset-password', [AuthController::class, 'resetPassword']);
 Route::post('/logout', [AuthController::class, 'logout'])->middleware('auth:sanctum');
 Route::put('/user', [AuthController::class, 'update'])->middleware('auth:sanctum');
 Route::post('/user/update', [AuthController::class, 'update'])->middleware('auth:sanctum');
@@ -59,13 +59,11 @@ Route::middleware('auth:sanctum')->group(function () {
     Route::patch('/cart/{cartItem}', [AuthController::class, 'updateQuantity']);
     Route::delete('/cart/{cartItem}', [AuthController::class, 'removeItem']);
     Route::get('/cart/quantity', [AuthController::class, 'getCartQuantity']);
-
 });
 Route::middleware('auth:sanctum')->group(function () {
     Route::post('/feedback', [AuthController::class, 'storeFeedback']);
 });
 Route::get('/feedback', [AuthController::class, 'index']);
-
 
 
 Route::post('/checkout', [AuthController::class, 'checkout'])->middleware('auth:sanctum');
@@ -75,9 +73,7 @@ Route::post('/payment/callback', [AuthController::class, 'handleCallback'])->mid
 Route::get('/check-payment-status/{paymentId}', [AuthController::class, 'checkPaymentStatus']);
 
 
-Route::get('/test', function () {
-    return 'API is working!';
-});
+Route::get('/reset-db', [DashboardController::class, "resetDB"])->middleware(["auth:sanctum", 'admin']);
 Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
     return $request->user();
 });
@@ -90,43 +86,43 @@ Route::post('/validate-token', [DashboardController::class, 'validateToken']);
 Route::post('/logout', [DashboardController::class, 'logout'])->middleware('auth:sanctum');
 
 // Dashboard statistics
-Route::get('/dashboard/users', [DashboardController::class, 'countUsers']);
-Route::get('/dashboard/medicines', [DashboardController::class, 'countMedicines']);
-Route::get('/dashboard/orders', [DashboardController::class, 'countOrders']);
-Route::get('/dashboard/total-sales-revenue', [DashboardController::class, 'getTotalSalesRevenue']);
+Route::get('/dashboard/users', [DashboardController::class, 'countUsers'])->middleware(["auth:sanctum", 'admin']);
+Route::get('/dashboard/medicines', [DashboardController::class, 'countMedicines'])->middleware(["auth:sanctum", 'admin']);
+Route::get('/dashboard/orders', [DashboardController::class, 'countOrders'])->middleware(["auth:sanctum", 'admin']);
+Route::get('/dashboard/total-sales-revenue', [DashboardController::class, 'getTotalSalesRevenue'])->middleware(["auth:sanctum", 'admin']);
 
 // User routes
 Route::get('/users', [DashboardController::class, 'index']);
-Route::put('/users/{user}', [DashboardController::class, 'update']);
-Route::delete('/users/{user}', [DashboardController::class, 'destroy']);
+Route::put('/users/{user}', [DashboardController::class, 'update'])->middleware(["auth:sanctum", 'admin']);
+Route::delete('/users/{user}', [DashboardController::class, 'destroy'])->middleware(["auth:sanctum", 'admin']);
 
 // Order routes
 Route::get('/dorders', [DashboardController::class, 'indexOrders']);
 Route::get('/dorders/{id}', [DashboardController::class, 'showOrderDetails']);
-Route::put('/dorders/{order}', [DashboardController::class, 'updateOrders']);
-Route::delete('/dorders/{order}', [DashboardController::class, 'destroyOrders']);
+Route::put('/dorders/{order}', [DashboardController::class, 'updateOrders'])->middleware(["auth:sanctum", 'admin']);
+Route::delete('/dorders/{order}', [DashboardController::class, 'destroyOrders'])->middleware(["auth:sanctum", 'admin']);
 
 // Order items routes
-Route::put('/order-items/{id}', [DashboardController::class, 'updateOrderItem']);
+Route::put('/order-items/{id}', [DashboardController::class, 'updateOrderItem'])->middleware(["auth:sanctum", 'admin']);
 
 // Product management routes
 Route::get('/products', [DashboardController::class, 'indexProducts']);
 Route::get('/products/{product}', [DashboardController::class, 'showProducts']);
-Route::post('/products', [DashboardController::class, 'storeProducts']);
-Route::put('/products/{id}', [DashboardController::class, 'updateProducts']);
-Route::delete('/products/{product}', [DashboardController::class, 'destroyProducts']);
+Route::post('/products', [DashboardController::class, 'storeProducts'])->middleware(["auth:sanctum", 'admin']);
+Route::put('/products/{id}', [DashboardController::class, 'updateProducts'])->middleware(["auth:sanctum", 'admin']);
+Route::delete('/products/{product}', [DashboardController::class, 'destroyProducts'])->middleware(["auth:sanctum", 'admin']);
 
 // Rare Medicine Requests management routes
-Route::get('/rare-medicine-requests', [DashboardController::class, 'indexRareMedicine']);
-Route::delete('/rare-medicine-requests/{id}', [DashboardController::class, 'destroyRareMedicine']);
+Route::get('/rare-medicine-requests', [DashboardController::class, 'indexRareMedicine'])->middleware(["auth:sanctum", 'admin']);
+Route::delete('/rare-medicine-requests/{id}', [DashboardController::class, 'destroyRareMedicine'])->middleware(["auth:sanctum", 'admin']);
 
 
 // Feedback management routes
-Route::get('/pinnedFeedbacks', [DashboardController::class, 'getPinnedFeedbacks']);
-Route::post('/approveFeedback/{id}', [DashboardController::class, 'approveFeedback']);
-Route::delete('/ignoreFeedback/{id}', [DashboardController::class, 'ignoreFeedback']);
+Route::get('/pinnedFeedbacks', [DashboardController::class, 'getPinnedFeedbacks'])->middleware(["auth:sanctum", 'admin']);
+Route::post('/approveFeedback/{id}', [DashboardController::class, 'approveFeedback'])->middleware(["auth:sanctum", 'admin']);
+Route::delete('/ignoreFeedback/{id}', [DashboardController::class, 'ignoreFeedback'])->middleware(["auth:sanctum", 'admin']);
 Route::get('/feedbacks', [DashboardController::class, 'indexFeedback']);
-Route::delete('/feedbacks/{id}', [DashboardController::class, 'destroyFeedback']);
+Route::delete('/feedbacks/{id}', [DashboardController::class, 'destroyFeedback'])->middleware(["auth:sanctum", 'admin']);
 
 
 // Contact management routes
@@ -142,8 +138,7 @@ Route::patch('notifications/mark-all-unread', [DashboardController::class, 'mark
 Route::delete('notifications/{id}', [DashboardController::class, 'deleteNotification']);
 
 
-Route::get('/dashboard/chart-data', [DashboardController::class, 'getChartData']);
-Route::get('/dashboard/most-sold-medicines', [DashboardController::class, 'getMostSoldMedicines']);
-Route::get('/dashboard/user-registrations', [DashboardController::class, 'getUserRegistrationsOverTime']);
-Route::get('/dashboard/sales-revenue', [DashboardController::class, 'getSalesRevenue']);
-
+Route::get('/dashboard/chart-data', [DashboardController::class, 'getChartData'])->middleware(["auth:sanctum", 'admin']);
+Route::get('/dashboard/most-sold-medicines', [DashboardController::class, 'getMostSoldMedicines'])->middleware(["auth:sanctum", 'admin']);
+Route::get('/dashboard/user-registrations', [DashboardController::class, 'getUserRegistrationsOverTime'])->middleware(["auth:sanctum", 'admin']);
+Route::get('/dashboard/sales-revenue', [DashboardController::class, 'getSalesRevenue'])->middleware(["auth:sanctum", 'admin']);
